Simple stress test and StressUtil API
